chore(dependencies): add BOM dependencies from igor and fiat
Of note in here is this moves the boot2 groovy version back up to 2.5.
We did a bad thing way back when we decided to use groovy-all everywhere instead of the
more specific groovy jars. The spring-boot-dependencies BOM defines versions for the
groovy component jars, but not groovy-all so we'll need to keep our groovy-all in
lock-step with the boot-dependencies groovy-all while we cut over to the smaller
groovy dependencies.