fix(exceptions): correctly log uncaught exceptions
While investigating another issue, I realized that uncaught thread exceptions aren't logged correctly.
They are logged to `stderr` and, depending on logstash config, may or may not show up in the logs.

This change configures a default uncaught thread exception handler and logs the exceptions as `error`s.
It then, optionally, passes the exception to the previous handler (if there was one).

This behavior can be disabled with `globalExceptionHandlingEnabled=false`