fix(resolvedEnv): unless explicitly configured, use Sanitizer defaults.
This was a regression from the boot2 upgrade where, now that there is a
ConfigurationProperties class for keysToSanitize, we are by default saying
do not sanitize anything.  This changes the default behaviour to use the
Sanitizer's default set of keywords to sanitize with.

While this is technically a breaking change, I think the expectation of
anyone using this would be that they don't have to supply configuration
to get a reasonable default level of resolvedEnv sanitiation, and anyone
who has supplied a custom value for this property will continue to work
as expected.