feat(plugins): Support configurable extensions
PF4J has two concepts of extension: System extensions and Plugin-provided extensions. This PR
refactors the current codebase to support configuring both concepts in a common way.

An example config might look something like this:

```yaml
spinnaker:
  plugins:
    robzienert.helloworld:
      enabled: true
      extensions:
        robzienert.helloworld-stage1:
          config:
            defaultRecipient: "Kork Reviewers"
        robzienert.helloworld-stage2:
          config:
            defaultRecipient: "Rob"
  extensions:
    robzienert.something-cool:
      enabled: true
      config:
        myConfig: myValue
```

The extension `$namespace.$id` is defined on the extension and is used to resolve config, the
same logic (just with a different pointer) to load extension config:

```kt
@SpinnakerExtension(namespace = "robzienert", id = "something-cool")
class SomethingCool : SomeSpinnakerExtensionPoint, ConfigurableExtension<SomethingCool.CoolConfig> {
  lateinit var config: CoolConfig

  fun setConfiguration(config: CoolConfig) {
    this.config = config
  }
}

class CoolConfig(
  val myConfig: String
)
```

A `ConfigResolver` is used to find configs and the default implementation is `SpringEnvironmentExtensionConfigResolver`.
This resolver will use the parent service Spring Environment `PropertySources` for all configs. Sourcing from other
places, like Front50, would be a matter of overriding the `ConfigResolver` - perhaps this becomes a resolver chain in
the future.

I had to refactor quite a lot of things to make this PR a thing, but I've moved stuff around and I feel the codebase
is easier to reason about as a result. One departure from stock PF4J is that all extensions are initialized once as
singletons.