fix(plugins): fix SpinnakerPluginManager initialization
Plugins didn't actually load. AbstractPluginManager conflates instantiation and initialization. The SpinnakerPluginManager pluginStatusProvider was delegating to null (even with the proxy in place) and load plugins, enabling and disabling failed with NPEs. The most straight forward way to fix it was to circumvent the create* methods and just initialize the protected members direct that are passing into SpinnakerPluginManager. There are tests that show the problem and fail without the changes.