refactor(artifacts): Deprecate Artifact constructors
The all-arg constructors for Artifact and ExpectedArtifact accept
a lot of similarly-typed parameters; to encourage use of the
Builder inside each of these classes, mark these constructors
deprecated with the goal of eventually removing them.

The no-arg constructors don't have a problem with too many parameters,
but they do encourage a pattern of creating an empty object and then
setting fields on it rather than creating a fully-formed object. (One
day it would be great to make these classes immutable, which would force
use of the Builder.) For this reason, also deprecate the no-arg
constructor.

Adding an annotation to a lombok-generated constructor is experimental
(and depends on Java version) so to be safe, I've just explicitly
defined the constructors for the purpose of adding the annotation.