feat(plugins): Add ability to download plugins with an external process
Prior to this PR, plugins can only be downloaded from unauthenticated HTTP
endpoints. This functionality adds the capability of specifying new
file downloader strategies, with an initial alternative strategy for
running an external process to download a plugin artifact.

We will be writing a new default FileDownloader that can be configured for
custom certificates, as well as basic auth, and probably other integrations
for artifact repositories that need something extra.