feat(plugins): plugin reference plugin type for local development (#466)
This introduces a new plugin type that is registered if pf4j is
set to development mode to enable a local plugin development
experience.

A plugin-ref is a JSON file with a pointer to the actual location
of a plugin and a list of the classes and libraries to include
when building the PluginClasspath.

This allows a runtime experience similar to dropping a fully
packaged plugin into the host application, without requiring
the packaging and deployment step - aside from a one time
generation and copy/link of the plugin-ref file.

This changeset includes all the plumbing to make plugin-ref
a supported type of plugin, but there is still work to do
in tooling and documentation to make the development
experience reasonable.

With this changeset and the following set of steps, you can
work with the host application and the plugin project open
in IntelliJ, launch the host application, and debug into
the plugin project. To do this you need to:

1. create a plugin-ref file, which is a JSON document:
```
{
  "pluginPath": "/path/to/plugin",
  "classesDirs": [
    "paths",
    "with",
    "code",
    "e.g. PLUGIN_PROJECT/build/classes/java/main"],
  "libsDirs": [
    "paths",
    "with",
    "jars"]
}
```

2. copy or link the plugin-ref file into the host
applications plugins directory

3. configure the host-application with the plugin enabled:
```
spinnaker.extensibility:
  plugins:
    <my plugin id>:
      enabled: true
```

4. run the host application with `pf4j.mode=development`

Given that set of steps you can then open the host
application project in IntelliJ. You can then add the
plugin project as a new project into the same workspace (
on the gradle tab, the `+` button).
Then create a Run Configuration for the host application,
but edit the Before Launch to do a Build Project instead
of just Build.

When you have alllll these steps complete, if you run the
host application you should see it pick up your plugin-ref,
and add your plugin. Additionally you can run in debug
mode and you will be able to set breakpoints in your plugin
and debug through it.