feat(annotations): Add annotations to default nullity
There is a ParametersAreNonnullByDefault annotation (and corresponding
nullable) in javax.annotation that has gained some use in Spinnaker's
microservices.  There is not, however, a corresponding annotation that
allows the default for fields or methods to be set.

Add annotations to specify that all fields within the scope are Nonnull
(or Nullable), and to specify that all methods within the scope reutrn
Nonnull (or Nullable).

Finally, add a NonnullByDefault annotation to specify that all methods,
fields, and parameters in the scope should be Nonnull unless explicitly
annotated (and a corresponding NullableByDefault).

The primary use case I envison for the Nullable annotations is to
reverse the effect of an annotation at a wider scope---perhaps an
entire package is annotated NonnullByDefault by there is a particular
class whose fields should all be nullable.