refactor(artifacts): Make ExpectedArtifact immutable
There was only ever one place where ExpectedArtifact was being
mutated, which was to set the boundArtifact. This place has been
updated to return a copy of the artifact instead of mutating it,
so none of the setters in this class are used. Remove the setters
and mark the fields (as well as the class) final.

Technically the class is only shallowly immutable, as Artifact itself
remains mutable, but this should still make it much easier to reason
about expected artifacts and reduce action-at-a-distance bugs.