feat(context): provide a RequestContext abstraction
This PR lifts Header out of AuthenticatedRequest and moves it into kork-core so that it can be accessed from both kork-web and kork-security.

In kork-web, we introduce a RequestContext interface with a concrete implementation that pulls values from AuthenticatedRequest. Consumers that need to access values from a RequestContext just need to get an autowired RequestContextProvider.

This should make code that depends on these values to be more easily testable and make it easier to go from one type of context to another (as I am currently doing in gate).