fix(ssl): pass TrustManager to okHttp SSL socketfactory as required by JDK9+
We recently encountered a crash loop using the igor 1.18.X codeline
when enabling SSL with a local truststore. The error being thrown is
```
Failed to instantiate [com.jakewharton.retrofit.Ok3Client]: Factory method
'ok3Client' threw exception; nested exception is java.lang.UnsupportedOperationException:
 clientBuilder.sslSocketFactory(SSLSocketFactory) not supported on JDK 9+
```
The stack trace allowed me to trace it back to kork's instantiation of the okHttpClient
which wasn't updated to pass in the TrustManager with the update of that JDK.

This only shows up if you are terminating SSL with your own truststore in the pod
so we may have been the first ones to come across this issue in the wild.