fix(telemetry): Prevent race in getMethodMetrics
It's possible to trigger a race in `getMethodMetrics`, specifically
when it's invoking `addInstrumentedMethod` for unprocessed metrics.

If multiple (tested with 20) pipelines are executed at the exact same time this
may trigger, if the execution would create new method metrics. In other words
it's quite rare, but I've managed to trigger it on my development instance
on a daily basis.