fix(artifacts): Improve null safety of expected artifact
The ExpectedArtifact class is annotated such that id and matchArtifact
are not null, but this is not actually enforced anywhere.

First, explicitly write out the AllArgsConstructor that was generated
by lombok, and move the Builder annotation to it (so the builder uses
that constructor when finally creating the object).

Then update the constructor to have a ParametersAreNullable annotation
as we can't control what gets passed to the constructor, and handle
when we are passed a null id or matchArtifact.

For the id case, we'll just translate to an empty string. For the
matchArtifact case, we'll create an empty matchArtifact. As the
point of a matchArtifact is to contain filters that determine whether
a given artifact matches this expected artifact; by defaulting to no
filters seems like a reasonable behavior here.