refactor(artifacts): Improve null-safety of Artifact class
The Artifact class contains a number of strings that are often null.
To make the class more null-safe, annotate it with NonnullByDefault and
translate any null values into the empty string in the constructor and
in setters.

In order to do this null -> empty translation, we need to explicitly
write out the constructor and the setters that were generated by
lombok.

We rarely mutate artifacts; when we do it often leads to confusing and
fragile logic because we often pass these many levels down in the call
stack, and also sometimes convert to/from Map<> meaning that it's not
always clear if you're operating on the same artifact or a copy. To
greatly simplify this, I'd like to make this class immutable. Based
on my analysis of where we mutate artifacts, the few places can all
be replaced by either (1) fully specifying the artifact up-front or
(2) returning a copy of the artifact with the desired mutations.

To start this process of making Artifact immutable, I've annotated
all of the setters as Deprecated. Once all of the places where we
call these are updated, I'll come back and actually remove the
setters.