refactor(artifacts): Revert null to empty conversion
In the last PR I was a bit overzealous in converting all null
entries in artifacts to emtpy string. This is a bit more invasive
than intended, as there are a number of places that are checking
only against null (and not empty). The benefit of this is not all
that clear given the potential to break things, so reverting that
part of the change.

The remainder of the PR remains---deprecating the setters.