fix(artifacts): Fix incorrect non-null annotation
Adding NullableByDefault to the Artifact class is also marking the
builder methods as Nullable which is incorrect, and annoying as
the generaly pattern is to chain calls. This is causing all calls
like Artifact.builder().name("abc").build() to have every call in
the chain be highlighted by editors that are aware of these
annotations.

Add a MethodsReturnNonnullByDefault to the builder class, as it will
always return non-null from its methods. (Also remove the
ParametersAreNullableByDefault that is implied by the top-level
annotation on the class.)