fix(artifacts): Fix non-null annotation
There's still one annoying thing that I didn't properly fix in the
last commit. Because the whole Artifact class is annotated as
NullableByDefault, we get a warning every time we call .toBuilder()
which *is* guaranteed to return non-null. We can't directly annotate
it as it's generated by Lombok; let's fix this by using
FieldsAreNullableByDefault so that it doesn't affect method values.

At first I wanted to affect method return values because the fact
that the fields are nullable means the accessors are also nullable.
But after testing, marking the fields as Nullable does cause the
lombok-generated accessors to be marked nullable, so marking the
fields nullable is enough.