refactor(artifacts): Add function getMetadata(string)
It's way safer to roll out making metadata shallowly immutable if
we remove the ability for consumers to get the raw map at all.

Currently consumer can call getMetadata and mutate the map; we can
make the map immutable but there's no safe way to do that with
warning to consumers. (They will go from getting a mutable map to
an immutable map and get a runtime error on mutating.)

A safer plan is to deprecate getMetadata in favor of having clients
ask for the metadata key of interest; once the function getMetadata
is removed then we will be able to transparently make the metadata
map immutable without affecting clients.

This of course is all just about making getMetadata shallowly
immutable. It unfortunately looks like there are some consumers that
are storing entire objects in metadata (in many cases just using an
artifact's metadata as a convenient place to store untyped data).
If I could start over, I would have make this a Map<String, String>
but it's too late for that, and we'll likely need to just live with
the fact that metadata is only shallowly immutable (and, what I think
is worse, can contain arbitrary untyped objects).