test(artifacts): Update artifact tests
Given that I've just updated a bunch of places where we are using
artifacts, now is a good time to add some tests on some of the
behavior consumers are expecting but that there are no tests for.

In particular:
* Improve the serialization/deserialization tests by separately
testing these two and by using JsonNodeFactory to create JSON
to test against.
* Add a lot more tests on the behavior of metadata as it's a bit
complex and worth explicitly checking that lombok/Jackson are
doing what we expect there. (Particularly to have test in palce
to make sure nothing breaks when we later remove the deprecated
methods.)
To ensure that toBuilder properly handles metadata, add a test.