refactor(artifacts): Shallow copy artifact metadata
As part of making Artifact immutable, we also want to make the
contents of metadata shallowly immutable.  (We can't enforce
deep immutability as it can contain arbitrary Objects.)

We've already deprecated (and removed from calling code) any way
for consumers to directly modify the metadata map (or even for them
to know that we are storing this as a Map).

That being said, there is still the loophole of clients passing in
a Map, then keeping around a reference and modifying it. Let's close
that loophole my making a shallow copy of the map before storing it on
the Artifact.

Also, remove an unecessary qualified reference to Artifact.ArtifactBuilder.