feat(dependencies): adds kork-runtime module
kork-runtime serves as a grouping of the runtime modules within kork for a
Spinnaker service (similar to a spring-boot-starter style module).

By collecting these opinions here, we can more easily refactor existing or
introduce new runtime opinions and propagate those to all the consuming
services.

This is specifically intended for modules that expose AutoConfiguration
or com.netflix.spinnaker.config style scanned Configuration classes and
are expected to be present in a running system, but aren't needed as a
compile time dependency.