feat(retrofit): add SpinnakerRetrofitExceptionHandlers and handle SpinnakerServerException and SpinnakerHttpException
SpinnakerRetrofitErrorHandler (introduced in https://github.com/spinnaker/kork/pull/821)
converts RetrofitErrors into SpinnakerServerException errors (or one of its subclasses),
but there was no corresponding exception handler for SpinnakerServerException.  So, the
generic handleException method from kork-web's GenericExceptionHandlers handled
SpinnakerServerException and logged a warning with a whole stack trace in addition to the
error.

So, this commit adds a handler for SpinnakerServerException and SpinnakerHttpException to
restore the pre-SpinnakerRetrofitErrorHandler behavior where GenericExceptionHandlers's
handler for RetrofitErrors extracted relevant information.  SpinnakerServerException is
now responsible for retrieving information from the underlying RetrofitError and storing
it in the exception's message.

It also adds a RetrofitErrorConfiguration class in the com.netflix.spinnaker.config
package, following the same convention as ErrorConfiguration in kork-web, so code that
includes kork-retrofit and scans that package gets a SpinnakerRetrofitExceptionHandlers
bean.