feat(artifacts): Add ArtifactStore to Kork
This commit adds a new feature called ArtifactStore. This allows for
storage of embedded/base64 types to some storage, eg S3. So instead of
having duplication of some artifact in a context, it'll now be
referenced by a artifact URI.

Also, there's a nice README that explains various settings as well as
how to setup a local S3 for testing artifact storage

[README link](kork-artifacts/src/main/java/com/netflix/spinnaker/kork/artifacts/README.md )

Signed-off-by: benjamin-j-powell <benjamin_j_powell@apple.com>