feat(crypto): Add X509IdentitySource API
This adds a new Kork module for common cryptographic operations. Primarily, this introduces the `X509IdentitySource` API for loading private keys and certificates as cryptographic identities and subsequently using them in `SSLContext` instances for configuring mutual TLS authentication. Included are two implementations for `X509Identity`: a static one containing a loaded private key and certificate chain, and a dynamic one which can be refreshed periodically to reload a key and certificate whenever the backing file changes.

Some other goodies that come with this new module:

- `SecureRandomBuilder` for creating a custom `SecureRandom` instance using a DRBG strategy (new in Java 9). This makes it simpler to configure domain-specific random number generators in a safe way.
- `PasswordProvider` interface for abstracting how the password for a keystore or private key is loaded (e.g., from a file).
- Updated version of Bouncycastle using the Java 1.8+ module.