feat(web): Add support for refreshable mTLS identity
This integrates the new kork-crypto module into kork-web to allow OkHttp v3 clients to use a refreshable identity. To avoid duplication of config logic and allowing this feature to be enabled via a feature flag, a new `OkHttpClientCustomizer` interface and prototype bean for `OkHttpClient.Builder` are introduced analogous to Spring's `WebClientCustomizer` and `WebClient.Builder` interfaces.

As a bonus, the `OkHttpClientComponents` config class is migrated from Groovy to Java (along with moving a little bit of Groovy annotations for better IDE detection).