feat(artifacts): Add ArtifactReferenceURI to artifact store
This commit adds a new URI type to the artifact store signature. This
allows for storage implementation to more easily use certain pieces of
the URI for storage. For example, with this commit, artifacts stored as
S3 objects will no longer have the `ref://` prefix, allowing for a
better AWS console experience

Signed-off-by: benjamin-j-powell <benjamin_j_powell@apple.com>