feat(artifacts): remove call to ArtifactStore.setInstance from ArtifactStore bean
Without this, spring tests that use ArtifactStoreConfiguration often fail.  More
specifically, because spring tests cache contexts (see
https://docs.spring.io/spring-framework/reference/testing/testcontext-framework/ctx-management/caching.html),
the spring context doesn't get destroyed, so ArtifactStore bean doesn't get destroyed, and
a second context then fails to start due to the IllegalStateException in
ArtifactStore.setInstance.

This puts the responsibility of calling ArtifactStore.setInstance on the code that
actually needs it.  In this case, that's orca because it uses ExpectedArtifact and
ExpressionsSupport.