feat(artifacts): Add HelmConfig and ensure properties aren't null
This commit adds the new HelmConfig which allows for users to turn on
expanding overrides. This will inspect overrides values, and if it is an
reference URI, it will expand and use that instead of the raw URI.

This also updates the properties, `helmConfig` and `s3` to no longer be
`null`.

Signed-off-by: benjamin-j-powell <bjp@apple.com>